Fixed packaging. Should now be able to properly run all tests and run the server
