+++ makerspace_backend/makerspace_backend_tests.py
@@ -14,5 +14,10 @@ class MakerSpaceBackendTestCase(unittest.TestCase):
         os.close(self.db_fd)
         os.unlink(makerspace_backend.app.config['DATABASE'])
 
+    def test_get_users(self):
+        resp = self.app.get('/v1/users')
+        for property, value in vars(resp).iteritems():
+            print property, ": ", value
+
 if __name__ == '__main__':
     unittest.main()
